Changed job_id to 32 bits per DPCS requirement. Jobs without job_id specified at submit time start with a value of 1<<16 and DPCS uses values 1 to 0xffffffff. This preclude effective use of a max job_id in search function, which was able to quickly indicate that a job_id requested by DPCS was not already in use. I added a job_id hash table to restore quick job find operations. Added check of packstr function to insure string would not go over 64k bytes, which is the limit for the protocol used.

+++ src/slurmctld/node_scheduler.c
@@ -43,7 +43,7 @@ int
 main (int argc, char *argv[]) 
 {
 	int error_code, error_count = 0, line_num, i;
-	uint16_t job_id;
+	uint32_t job_id;
 	FILE *command_file;
 	char in_line[BUF_SIZE], *node_list;
 	log_options_t opts = LOG_OPTS_STDERR_ONLY;
